rangy as UMD module and reorganize rangy-classapplier plugin to the correct folder
+ Microsoft/TypeScript#7125
+ rangy-classapplier.js is in [lib folder](https://github.com/timdown/rangy/tree/master/lib)
