+++ types/rangy/tsconfig.json
@@ -19,7 +19,7 @@
     },
     "files": [
         "index.d.ts",
-        "rangy-classapplier.d.ts",
+        "lib/rangy-classapplier.d.ts",
         "rangy-tests.ts"
     ]
 }
